requirements.txt
climage
google-genai
rich